refactor edit function
